Don't crash if media pool directory does not exist and immutable suites are used

+++ source/datacache.d
@@ -503,6 +503,9 @@ public:
 
         void cleanupDirs (string rootPath) {
             auto pdir = buildNormalizedPath (rootPath, "..");
+            if (!std.file.exists (pdir))
+                return;
+
             if (dirEmpty (pdir))
                 rmdir (pdir);
             pdir = buildNormalizedPath (pdir, "..");
